3DPlot initial implementation
Added 3DPlot objects. The new feature is optional, it requires
QtDataVisualization module. The 3DPlot was primarily intended to be
used for short-time Fourier transform visualization, but can also be
used as a generic 3D plotting tool. It is much more efficient than
Octave's 3D plot when it comes to large data sets, but not as universal.
For now it can only display 3D surfaces. The UI has some minor quirks
and limitations related to QtDataVisualization implementation.